pyproject.toml
[project]
name = "code-revieww-with-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "ollama>=0.4.7",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]
